Move revert test to own file
